Rework Server Startup to Include Resources
Reworks the entire server startup to include resource instantiation, and
changes the way the Server is initialized. Drops most of the heavy
lifting of building and initializing the resources down into a few
builders, and packages the resource proxies together into a single
struct that can be passed around and cloned.